Moved buttons from hats to buttons class
