+++ neo_batterylevelshutdown/page_exit.py
@@ -10,20 +10,10 @@
 ===========================================
 """
 
-from datetime import datetime
 import os.path
-import subprocess
 import sys
-from PIL import Image, ImageFont, ImageDraw
+from PIL import Image
 from .HAT_Utilities import get_device
-
-
-try:
-    import psutil
-except ImportError:
-    print("The psutil library was not found. "
-          "Run 'sudo -H pip install psutil' to install it.")
-    sys.exit()
 
 
 class PageExit:
