+++ neo_batterylevelshutdown/usb.py
@@ -3,7 +3,6 @@
 import logging
 import os
 import subprocess
-import threading
 from distutils.dir_util import copy_tree
 
 
@@ -13,7 +12,7 @@ class USB:
     def __init__(self):
         pass
 
-    def isUsbPresent(self, devPath = '/dev/sda1'):
+    def isUsbPresent(self, devPath='/dev/sda1'):
         '''
         Returns if there is a USB plugged into specified devPath
         :return: True / False
@@ -22,7 +21,7 @@ class USB:
         return os.path.exists(devPath)
 
 
-    def unmount(self, curPath = '/media/usb0'):
+    def unmount(self, curPath='/media/usb0'):
         '''
         Unmount the USB drive from curPath
         :return:  True / False
@@ -34,7 +33,7 @@ class USB:
         except:
             return False
 
-    def mount(self, devPath = '/dev/sda1', newPath = '/media/usb1'):
+    def mount(self, devPath='/dev/sda1', newPath='/media/usb1'):
         '''
         Mount the USB drive at the devPath to the specified newPath location
 
@@ -44,14 +43,15 @@ class USB:
         # try:
         logging.debug("Mounting USB at {} to {}".format(devPath, newPath))
         if not os.path.exists(newPath):  # see if desired mounting directory exists
-            os.makedirs(newPath)  # if not, make it, and all of the intermediary directories if needed
-        response = subprocess.call(['mount','-o', 'sync,noexec,nodev,noatime,nodiratime,utf8', devPath, newPath])
+            os.makedirs(newPath)  # create directory and all of the intermediary directories
+        response = subprocess.call(['mount', '-o', 'sync,noexec,nodev,noatime,nodiratime,utf8',
+                                    devPath, newPath])
         logging.debug("Response: {}".format(response))
         return True if response == 0 else False
         # except:
         #     return False
 
-    def copyFiles(self, sourcePath = '/media/usb1', destPath = '/media/usb0'):
+    def copyFiles(self, sourcePath='/media/usb1', destPath='/media/usb0'):
         '''
         Move files from sourcePath to destPath recursively
         :param sourcePath: place where files are
@@ -70,7 +70,7 @@ class USB:
         else:
             return False
 
-    def checkSpace(self, sourcePath = '/media/usb1', destPath = '/media/usb0'):
+    def checkSpace(self, sourcePath='/media/usb1', destPath='/media/usb0'):
         '''
         Function to make sure there is space on destination for source materials
 
@@ -81,11 +81,9 @@ class USB:
         if os.path.exists(sourcePath) and os.path.exists(destPath):
             sourceSize = self.getSize(sourcePath)
             destSize = self.getFreeSpace(destPath)
-            logging.debug("Source size: {} bytes, destination size: {} bytes".format(sourceSize, destSize))
-            if destSize >= sourceSize :
-                return True
-            else:
-                return False
+            logging.debug("Source size: {} bytes, destination size: {} bytes".format(
+                sourceSize, destSize))
+            return bool(destSize >= sourceSize)
         else:
             return False
 
@@ -118,11 +116,12 @@ class USB:
         adjustedFree = free - freeSpaceCushion
         return adjustedFree
 
-    def moveMount(self, devMount = '/dev/sda1', curMount = '/media/usb0', destMount = '/media/usb1'):
+    def moveMount(self, devMount='/dev/sda1', curMount='/media/usb0', destMount='/media/usb1'):
         '''
         This is a wrapper for umount, mount.  This is simple and works.
-        we could use mount --move  if the mount points are not within a mount point that is marked as shared,
-        but we need to consider the implications of non-shared mounts before doing it
+        we could use mount --move  if the mount points are not within a mount point that is
+        marked as shared, but we need to consider the implications of non-shared mounts before
+        doing it
 
         :param devMount: device name in the /dev listing
         :param curMount: where usb is currently mounted
